requirements.txt
Pillow==8.4.0
pytz==2021.3
speedtest-cli==2.1.3
streamlit==1.8.1
streamlit-option-menu==0.3.2